Added all the materials for float precision scan_fov.

+++ headers/program.h
@@ -6,7 +6,7 @@
 /*   By: jnivala <jnivala@student.hive.fi>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/22 13:39:02 by tmaarela          #+#    #+#             */
-/*   Updated: 2021/02/19 14:37:34 by jnivala          ###   ########.fr       */
+/*   Updated: 2021/02/19 16:52:29 by jnivala          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -74,7 +74,7 @@ typedef struct		s_frame
 	int				idx;
 	float			max_fov;
 	float			offset;
-	float			min_step;
+	double			min_step;
 	int				old_idx;
 	SDL_Surface		*draw_surf;
 	t_xy			plr_offset;
